+++ NEWS.md
@@ -10,7 +10,6 @@ funcy version 1.0.1
     * "distFPCA" was changed to "distclust". This was a bug in the help file.
     * "funclust" was removed from examples since its depends on R package "Funclustering" which has been archieved. 
 
-
 * funcyOUtList.Rd
     * "funclust" was removed from examples since its depends on R package "Funclustering" which has been archieved. 
 
@@ -20,3 +19,8 @@ funcy version 1.0.1
 * test-old-unit-test.R
     * "funclust" was removed from unit tests since its depends on R package "Funclustering" which has been archieved. 
 
+* xecute.R
+    * warning for methods = "funclust" saying that funclust is currently not available on CRAN. 
+
+* DESCRIPTION
+    * Funclustering was removed from suggests .